MyEventSystem Class fix
I deleted the assembly definition file in the script folder. I tried adding the override reference to the GameAnalytics.dll file but I didn't find a way to make it show up in the dropdown menu. So, as the asmdef file is only compiling the MyEventSystem.cs file, it really doesn't make a difference if it's there or not, so I chose to delete it and the game compiles.