Level Management
I wrote a few lines in the project so that I can get the level index and go to the next level without having to write the number in the inspector of each object in every level. This way, we only need to indicate the last level and it will loop to the first level by itself.